pyproject.toml
[tool.poetry]
name = "visdom-scraper"
version = "0.1.0"
description = "A tool to scrape websites and convert them to markdown"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"       # Updated to support markitdown
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
scrapy = "^2.11.0"
selenium = "^4.15.2"
webdriver-manager = "^4.0.1"
urllib3 = "^2.0.7"
click = "^8.1.7"
python-slugify = "^8.0.1"
colorlog = "^6.7.0"
html2text = "^2020.1.16"
markitdown = "^0.1.1"        # Added markitdown as a direct dependency

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
visdom-scraper = "visdom_scraper.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
